requirements.txt
# 核心GIS库 
geopandas>=0.14.0
rasterio>=1.3.0
pyproj>=3.6.0
shapely>=2.0.0
fiona>=1.9.0
# gdal>=3.8.0  # 建议先注释掉，pip直接装GDAL源码极大概率会编译失败

# 空间分析与建模
scikit-learn>=1.3.0
scipy>=1.11.0
numpy>=1.24.0
pandas>=2.0.0
xarray>=2023.6.0

# 生态学专用库 & 跨语言方案
pysal>=2.7.0
juliacall          # 用于跨语言桥接 Julia 版的 Circuitscape
pylandstats        # 替代 R 语言 landscapemetrics 的纯 Python 景观格局方案
rpy2               # 若仍需调用 R 语言生态库可保留此项

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
folium>=0.14.0
contextily>=1.5.0

# 工具类
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0
scikit-image>=0.21.0
opencv-python      # 预留用于处理红外相机图像提取

# 报告生成
jinja2>=3.1.0
openpyxl>=3.1.0
pdfkit